backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
python-multipart
pyjwt
passlib[bcrypt]
python-dotenv
